agentic_rag/requirements.txt
Crawl4AI
python-dotenv
certifi
pyaml
google-generativeai
openai
supabase
pydantic-ai
logfire
langgraph
langsmith
langchain_anthropic
langchain-google-genai
streamlit
PyPDF2
python-docx
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-parse
llama-index-extractors-entity
llama-index-postprocessor-cohere-rerank
fs

faiss-cpu